Made files more uniform
