+++ src/app/app.routes.ts
@@ -1,15 +1,16 @@
-import { Routes } from '@angular/router';
-import { AuthorisationComponent } from './authorisation/authorisation.component';
-import { LoginComponent } from './authorisation/login/login.component';
-import { NewUserComponent } from './authorisation/new-user/new-user.component';
-import { HomeComponent } from './home/home.component';
-import { UserComponent } from './user/user.component';
-import { DashboardComponent } from './dashboard/dashboard.component';
-import { AngularFireAuthGuard,redirectLoggedInTo,redirectUnauthorizedTo } from '@angular/fire/compat/auth-guard';
-import { ProfileComponent } from './user/profile/profile.component';
-import { TransactionsViewComponent } from './transactions-view/transactions-view.component';
-import { accountCreationGuard } from './authorisation.service';
-import { accountResolver } from './shared/resolvers/account.resolver';
+import { Routes }												from '@angular/router';
+import { AuthGuard, redirectLoggedInTo,redirectUnauthorizedTo }	from '@angular/fire/auth-guard';
+
+import { AuthorisationComponent }		from './authorisation/authorisation.component';
+import { accountCreationGuard }			from './shared/guards/create-account.guard';
+import { accountResolver }				from './shared/resolvers/account.resolver';
+import { DashboardComponent }			from './dashboard/dashboard.component';
+import { HomeComponent }				from './home/home.component';
+import { LoginComponent }				from './authorisation/login/login.component';
+import { NewUserComponent }				from './authorisation/new-user/new-user.component';
+import { ProfileComponent }				from './user/components/profile/profile.component';
+import { TransactionsViewComponent} 	from './transactions-view/transactions-view.component';
+import { UserComponent }				from './user/user.component';
 
 const redirectUnauthorizedToLogin = () => redirectUnauthorizedTo(['account', 'login']);
 const redirectAuthorisedToDashboard = () => redirectLoggedInTo(['dashboard']);
@@ -18,7 +19,7 @@ export const routes: Routes = [
 	{
 		path: '',
 		component: HomeComponent,
-		canActivate: [AngularFireAuthGuard],
+		canActivate: [AuthGuard],
 		data: { authGuardPipe: redirectAuthorisedToDashboard },
   	},
   	{
@@ -32,7 +33,7 @@ export const routes: Routes = [
   	{
 		path: 'user',
 		component: UserComponent,
-		canActivate: [AngularFireAuthGuard],
+		canActivate: [AuthGuard],
 		resolve: { data: accountResolver },
 		data: { authGuardPipe: redirectUnauthorizedToLogin },
 		children: [{ path: 'profile', component: ProfileComponent }],
@@ -41,13 +42,13 @@ export const routes: Routes = [
 		path: 'dashboard',
 		component: DashboardComponent,
 		resolve: { data: accountResolver },
-		canActivate: [AngularFireAuthGuard],
+		canActivate: [AuthGuard],
 		data: { authGuardPipe: redirectUnauthorizedToLogin },
 	},
 	{
 		path: 'transactions',
 		component: TransactionsViewComponent,
-		canActivate: [AngularFireAuthGuard],
+		canActivate: [AuthGuard],
 		resolve: {data: accountResolver},
 		data: { authGuardPipe: redirectUnauthorizedToLogin },
 	},
