--- src/app/shared/amount.ts
@@ -1,5 +1,0 @@
-
-export interface Amount {
-  name: string;
-  amount: number;
-}
