+++ src/app/shared/components/transactions-table/transactions-table.component.css
