+++ src/app/shared/interfaces/status.interface.ts
