+++ src/app/shared/services/accounts.service.ts
@@ -1,116 +1,112 @@
-import { Injectable } from '@angular/core';
-import { collection,Firestore,collectionData,DocumentData,addDoc,deleteDoc,doc } from '@angular/fire/firestore';
-import { BehaviorSubject, map, take } from 'rxjs';
-import { AuthorisationService } from 'src/app/authorisation.service';
-import { Account, DraftAccount } from 'src/app/user/account/account.interface';
+import { Injectable }																						from '@angular/core';
+import { addDoc, collection, collectionData, CollectionReference, deleteDoc, doc, DocumentData, Firestore }	from '@angular/fire/firestore';
+import { BehaviorSubject, map, take }																		from 'rxjs';
+
+import { AuthorisationService }		from 'src/app/shared/services/authorisation.service';
+import { Account, DraftAccount }	from 'src/app/shared/interfaces/account.interface';
 
 @Injectable({
-  providedIn: 'root',
+	providedIn: 'root',
 })
 export class AccountsService {
-  public _accounts: Account[] = [];
-  public accounts$: BehaviorSubject<Account[]> = new BehaviorSubject<Account[]>([]);
+	private _accounts: Account[] = [];
+	public accounts$: BehaviorSubject<Account[]> = new BehaviorSubject<Account[]>([]);
 
+	private accountsPath: string;
+	private accountsCollection: CollectionReference;
 
+	constructor(private fireStore: Firestore, private auth: AuthorisationService) {
+		this.accountsPath = `users/${this.auth.getUserId()}/Accounts`;
+		this.accountsCollection = collection(
+			this.fireStore,
+			this.accountsPath,
+		);
+	}
 
-  constructor(private fireStore: Firestore, private auth: AuthorisationService) {}
+	/* -------------------------------------------------------------------------- */
+	/*                              Firebase Fetching                             */
+	/* -------------------------------------------------------------------------- */
+	// Retrieve accounts from firebase and assign to local data
+	public fetchAccounts(): void {
+		collectionData(this.accountsCollection, { idField: 'id' })
+			.pipe(
+				take(1),
+				map((accounts: DocumentData | (DocumentData & {})) => {
+					let returnedAccounts: Account[] = [];
+					accounts.forEach((account: Account) => {
+						returnedAccounts.push(account);
+					});
+					return returnedAccounts;
+				}),
+			)
+			.subscribe((accounts: Account[]) => {
+				this.setAccountsSubject(accounts);
+			});
+	}
 
-  /* -------------------------------------------------------------------------- */
-  /*                              Firebase Fetching                             */
-  /* -------------------------------------------------------------------------- */
-  // Retrieve accounts from firebase and assign to local data
-  public fetchAccounts(): void {
-    const accountsCollection = collection(
-      this.fireStore,
-      'users/' + this.auth.getUserId() + '/Accounts',
-    );
+	/* -------------------------------------------------------------------------- */
+	/*                                   Create                                   */
+	/* -------------------------------------------------------------------------- */
+	public addAccount(newAccount: DraftAccount): void {
+		addDoc(this.accountsCollection, newAccount).then((response) => {
+			// add to list of accounts
+			const createdAccount: Account = { ...newAccount, id: response.id };
+			let updatedAccountList: Account[] = [...this._accounts, createdAccount];
+			this.setAccountsSubject(updatedAccountList);
 
-    collectionData(accountsCollection, { idField: 'id' })
-      .pipe(
-        take(1),
-        map((accounts: DocumentData | (DocumentData & {})) => {
-          let returnedAccounts: Account[] = [];
-          accounts.forEach((account: Account) => {
-            returnedAccounts.push(account);
-          });
-          return returnedAccounts;
-        }),
-      )
-      .subscribe((accounts: Account[]) => {
-        this.setAccountsSubject(accounts);
-      });
-  }
+			if(createdAccount.type === 'Savings') {
+				this.addSavingsPot(createdAccount, 'Main', createdAccount.amount);
+			}
+		});
+	}
 
-  /* -------------------------------------------------------------------------- */
-  /*                                   Create                                   */
-  /* -------------------------------------------------------------------------- */
-  public addAccount(newAccount: DraftAccount): void {
-    const accountsCollection = collection(
-      this.fireStore,
-      'users/' + this.auth.getUserId() + '/Accounts',
-    );
-    addDoc(accountsCollection, newAccount).then((response) => {
-      // add to list of accounts
-      const createdAccount: Account = { ...newAccount, id: response.id };
-      let updatedAccountList: Account[] = [...this._accounts, createdAccount];
-      this.setAccountsSubject(updatedAccountList);
+	// Add a pot to an existing account
+	public addSavingsPot(account: Account, name: string, amount: number): void {
+		const potsPath = `${this.auth.getUserId()}/Accounts/${account.id}/pots`
+		const potCollection = collection(
+			this.fireStore,
+			potsPath,
+		);
+		addDoc(potCollection, { name: name, amount: amount ?? account.amount });
+	}
 
-      if(createdAccount.type === 'Savings') {
-        this.addSavingsPot(createdAccount, 'Main', createdAccount.amount);
-      }
-    });
-  }
+	/* -------------------------------------------------------------------------- */
+	/*                                    Read                                    */
+	/* -------------------------------------------------------------------------- */
+	public getAccount(accountId: string): Account {
+		const account: Account | undefined = this._accounts.find((account: Account) => account.id === accountId);
+		if(account) return account;
+		throw new Error('Invalid account Id');
+	}
 
-  // Add a pot to an existing account
-  public addSavingsPot(account: Account, name: string, amount: number): void {
-    const potCollection = collection(
-      this.fireStore,
-      'users/' + this.auth.getUserId() + '/Accounts/' + account.id + '/pots',
-    );
-    addDoc(potCollection, { name: name, amount: amount ?? account.amount });
-  }
+	/* -------------------------------------------------------------------------- */
+	/*                                   Update                                   */
+	/* -------------------------------------------------------------------------- */
 
-  /* -------------------------------------------------------------------------- */
-  /*                                    Read                                    */
-  /* -------------------------------------------------------------------------- */
-  public getAccount(accountId: string): Account {
-    const account: Account | undefined = this._accounts.find((account: Account) => account.id === accountId);
-    if(account) return account;
-    throw 'Invalid account Id';
-  }
+	/* -------------------------------------------------------------------------- */
+	/*                                   Delete                                   */
+	/* -------------------------------------------------------------------------- */
+	public deleteAccount(accountId: string): void {
+		deleteDoc(doc(this.accountsCollection, accountId)).then((response) => {
+			let updatedAccountList = this._accounts.filter(
+				(account: Account) => account.id !== accountId,
+			);
 
-  /* -------------------------------------------------------------------------- */
-  /*                                   Update                                   */
-  /* -------------------------------------------------------------------------- */
+			this.setAccountsSubject(updatedAccountList);
+		});
+	}
 
-  /* -------------------------------------------------------------------------- */
-  /*                                   Delete                                   */
-  /* -------------------------------------------------------------------------- */
-  public deleteAccount(accountId: string): void {
-    const accountsCollection = collection(
-      this.fireStore,
-      'users/' + this.auth.getUserId() + '/Accounts',
-    );
-    deleteDoc(doc(accountsCollection, accountId)).then((response) => {
-      let updatedAccountList = this._accounts.filter(
-        (account: Account) => account.id !== accountId,
-      );
+	/* -------------------------------------------------------------------------- */
+	/*                                Miscellaneous                               */
+	/* -------------------------------------------------------------------------- */
+	// Check if the account data has been fetched from firebase
+	public hasAccounts(): boolean {
+		return !!this._accounts.length;
+	}
 
-      this.setAccountsSubject(updatedAccountList);
-    });
-  }
-
-  /* -------------------------------------------------------------------------- */
-  /*                                Miscellaneous                               */
-  /* -------------------------------------------------------------------------- */
-  // Check if the account data has been fetched from firebase
-  public hasAccounts(): boolean {
-    return !!this._accounts.length;
-  }
-
-  // Set the local account variable and observable
-  private setAccountsSubject(accounts: Account[]): void {
-    this._accounts = accounts;
-    this.accounts$.next(this._accounts);
-  }
+	// Set the local account variable and observable
+	private setAccountsSubject(accounts: Account[]): void {
+		this._accounts = accounts;
+		this.accounts$.next(this._accounts);
+	}
 }
