+++ src/app/user/components/add-account/add-account.component.html
@@ -1,0 +1,24 @@
+<h3 mat-dialog-title>{{ action }}</h3>
+<div mat-dialog-content>
+    <form (ngSubmit)="submitAccount()" [formGroup]="accountForm">
+        <mat-form-field>
+            <mat-label>Account Name</mat-label>
+            <input type="text" matInput formControlName="name" />
+        </mat-form-field>
+        <mat-form-field>
+            <mat-label>Type of Account</mat-label>
+            <mat-select formControlName="type">
+                @for (option of accountTypes; track $index) {
+                    <mat-option [value]="option">{{ option }}</mat-option>
+                }
+            </mat-select>
+        </mat-form-field>
+        @if (showNum) {
+            <mat-form-field>
+                <mat-label>Amount</mat-label>
+                <input type="number" matInput formControlName="amount" />
+            </mat-form-field>
+        }
+        <button mat-raised-button>Submit</button>
+    </form>
+</div>
