+++ src/app/user/user.component.ts
@@ -1,15 +1,15 @@
-import { Component, OnInit } from '@angular/core';
-import { RouterOutlet } from '@angular/router';
+import { Component }  	from '@angular/core';
+import { RouterOutlet }	from '@angular/router';
 
 @Component({
-    selector: 'app-user',
-    templateUrl: './user.component.html',
-    styleUrls: ['./user.component.css'],
-    standalone: true,
-    imports: [RouterOutlet]
+	selector: 'app-user',
+	templateUrl: './user.component.html',
+	styleUrls: ['./user.component.css'],
+	standalone: true,
+	imports: [RouterOutlet]
 })
 export class UserComponent {
 
-  constructor() {}
+	constructor() {}
 
 }
