+++ src/test/accounts.service.spec.ts
@@ -2,7 +2,7 @@ import { TestBed } from '@angular/core/testing';
 
 import { AccountsService } from '../app/shared/accounts.service';
 
-import { AuthorisationService } from 'src/app/authorisation.service';
+import { AuthorisationService } from 'src/app/shared/services/authorisation.service';
 import { Firestore } from '@angular/fire/firestore';
 
 describe('AccountsService', () => {
